Consolidate small JS engines and clean up legacy files
Consolidate small engines (<10KB) into `js/engines/` bundles (ui, data, engagement, booking-utils).
Update JS wrappers to use lazy-loaded bundles.
Refactor `js/payment.js` to use `loadDeferredModule` for payment logic.
Merge legacy lazy-loading scripts (`gallery-lazy.js`, `booking-calendar-lazy-init.js`) into `js/main.js`.
Delete obsolete files.
Update `rollup.config.mjs` to remove unnecessary chunks.
Verify frontend functionality with Playwright.
