--- booking-calendar-lazy-init.js
@@ -1,28 +1,0 @@
-(function () {
-    'use strict';
-
-    function wireBookingCalendarLazyLoad(element) {
-        if (!element) {
-            return;
-        }
-
-        element.addEventListener('click', function () {
-            import('./js/booking-calendar.js').then(function (moduleRef) {
-                if (moduleRef && typeof moduleRef.initCalendar === 'function') {
-                    moduleRef.initCalendar();
-                }
-            }).catch(function () {
-                // noop
-            });
-        });
-    }
-
-    const bookingBtn = document.getElementById('booking-btn');
-    wireBookingCalendarLazyLoad(bookingBtn);
-
-    document.querySelectorAll('a[href="#citas"]').forEach(function (button) {
-        if (button.id !== 'booking-btn') {
-            wireBookingCalendarLazyLoad(button);
-        }
-    });
-})();
