+++ booking-ui.js
@@ -103,15 +103,14 @@
 
         async function updateAvailableTimes() {
             try {
-                const module = await import('./js/booking-calendar.js');
-                await module.updateAvailableTimes(deps, {
+                await deps.updateAvailableTimes({
                     dateInput,
                     timeSelect,
                     doctorSelect,
                     t
                 });
             } catch (error) {
-                console.error('Failed to load booking-calendar.js', error);
+                console.error('Failed to load booking-calendar', error);
                 deps.showToast(t('Error cargando calendario. Intenta nuevamente.', 'Error loading calendar. Please try again.'), 'error');
             }
         }
