+++ index.html
@@ -1932,8 +1932,6 @@
     </div>
     <script src="bootstrap-inline-engine.js?v=figo-bootstrap-20260220-cachecoherence1" defer></script>
     <script src="script.js?v=figo-20260220-phase6-cachecoherence1" defer></script>
-    <script src="gallery-lazy.js" defer></script>
-    <script src="booking-calendar-lazy-init.js" defer></script>
 </body>
 </html>
