+++ js/engagement.js
@@ -3,8 +3,7 @@ import { loadDeferredModule, runDeferredModule, withDeferredModule, createWarmup
 import { getCurrentLang, getReviewsCache, setReviewsCache } from './state.js';
 import { apiRequest, createCallbackRecord, createReviewRecord } from './data.js';
 
-const ENGAGEMENT_FORMS_ENGINE_URL = withDeployAssetVersion('/engagement-forms-engine.js?v=figo-engagement-20260218-phase1-sync1');
-const REVIEWS_ENGINE_URL = withDeployAssetVersion('/reviews-engine.js?v=figo-reviews-20260219-phase1');
+const ENGAGEMENT_BUNDLE_URL = withDeployAssetVersion('/js/engines/engagement-bundle.js');
 
 // REVIEWS ENGINE
 function getReviewsEngineDeps() {
@@ -18,14 +17,14 @@ function getReviewsEngineDeps() {
 
 export function loadReviewsEngine() {
     return loadDeferredModule({
-        cacheKey: 'reviews-engine',
-        src: REVIEWS_ENGINE_URL,
-        scriptDataAttribute: 'data-reviews-engine',
+        cacheKey: 'engagement-bundle',
+        src: ENGAGEMENT_BUNDLE_URL,
+        scriptDataAttribute: 'data-engagement-bundle',
         resolveModule: () => window.PielReviewsEngine,
         isModuleReady: (module) => !!(module && typeof module.init === 'function'),
         onModuleReady: (module) => module.init(getReviewsEngineDeps()),
         missingApiError: 'reviews-engine loaded without API',
-        loadError: 'No se pudo cargar reviews-engine.js',
+        loadError: 'No se pudo cargar reviews-engine (engagement-bundle)',
         logLabel: 'Reviews engine'
     });
 }
@@ -66,17 +65,17 @@ function getEngagementFormsEngineDeps() {
 }
 
 export function loadEngagementFormsEngine() {
-    return loadReviewsEngine().then(() => loadDeferredModule({
-        cacheKey: 'engagement-forms-engine',
-        src: ENGAGEMENT_FORMS_ENGINE_URL,
-        scriptDataAttribute: 'data-engagement-forms-engine',
+    return loadDeferredModule({
+        cacheKey: 'engagement-bundle',
+        src: ENGAGEMENT_BUNDLE_URL,
+        scriptDataAttribute: 'data-engagement-bundle',
         resolveModule: () => window.PielEngagementFormsEngine,
         isModuleReady: (module) => !!(module && typeof module.init === 'function'),
         onModuleReady: (module) => module.init(getEngagementFormsEngineDeps()),
         missingApiError: 'engagement-forms-engine loaded without API',
-        loadError: 'No se pudo cargar engagement-forms-engine.js',
+        loadError: 'No se pudo cargar engagement-forms-engine (engagement-bundle)',
         logLabel: 'Engagement forms engine'
-    }));
+    });
 }
 
 export function initEngagementFormsEngineWarmup() {
