+++ js/reschedule.js
@@ -4,11 +4,10 @@ import { getCurrentLang } from './state.js';
 import { DEFAULT_TIME_SLOTS } from './config.js';
 import { apiRequest, loadAvailabilityData, getBookedSlots, invalidateBookedSlotsCache } from './data.js';
 
-const RESCHEDULE_GATEWAY_ENGINE_URL = withDeployAssetVersion('/reschedule-gateway-engine.js?v=figo-reschedule-gateway-20260219-phase1');
+const BOOKING_UTILS_URL = withDeployAssetVersion('/js/engines/booking-utils.js');
 
-function getRescheduleGatewayEngineDeps() {
+function getRescheduleEngineDeps() {
     return {
-        loadDeferredModule,
         apiRequest,
         loadAvailabilityData,
         getBookedSlots,
@@ -20,23 +19,23 @@ function getRescheduleGatewayEngineDeps() {
     };
 }
 
-export function loadRescheduleGatewayEngine() {
+export function loadRescheduleEngine() {
     return loadDeferredModule({
-        cacheKey: 'reschedule-gateway-engine',
-        src: RESCHEDULE_GATEWAY_ENGINE_URL,
-        scriptDataAttribute: 'data-reschedule-gateway-engine',
-        resolveModule: () => window.PielRescheduleGatewayEngine,
+        cacheKey: 'booking-utils',
+        src: BOOKING_UTILS_URL,
+        scriptDataAttribute: 'data-booking-utils',
+        resolveModule: () => window.PielRescheduleEngine,
         isModuleReady: (module) => !!(module && typeof module.init === 'function'),
-        onModuleReady: (module) => module.init(getRescheduleGatewayEngineDeps()),
-        missingApiError: 'reschedule-gateway-engine loaded without API',
-        loadError: 'No se pudo cargar reschedule-gateway-engine.js',
-        logLabel: 'Reschedule gateway engine'
+        onModuleReady: (module) => module.init(getRescheduleEngineDeps()),
+        missingApiError: 'reschedule-engine loaded without API',
+        loadError: 'No se pudo cargar reschedule-engine (booking-utils)',
+        logLabel: 'Reschedule engine'
     });
 }
 
 export function initRescheduleEngineWarmup() {
     runDeferredModule(
-        loadRescheduleGatewayEngine,
+        loadRescheduleEngine,
         (engine) => engine.initRescheduleFromParam(),
         () => {
             showToast(getCurrentLang() === 'es' ? 'No se pudo cargar la reprogramacion.' : 'Unable to load reschedule flow.', 'error');
@@ -45,7 +44,7 @@ export function initRescheduleEngineWarmup() {
 }
 
 export function closeRescheduleModal() {
-    runDeferredModule(loadRescheduleGatewayEngine, (engine) => engine.closeRescheduleModal(), () => {
+    runDeferredModule(loadRescheduleEngine, (engine) => engine.closeRescheduleModal(), () => {
         const modal = document.getElementById('rescheduleModal');
         if (modal) {
             modal.classList.remove('active');
@@ -54,7 +53,7 @@ export function closeRescheduleModal() {
 }
 
 export function submitReschedule() {
-    runDeferredModule(loadRescheduleGatewayEngine, (engine) => engine.submitReschedule(), () => {
+    runDeferredModule(loadRescheduleEngine, (engine) => engine.submitReschedule(), () => {
         showToast(getCurrentLang() === 'es' ? 'No se pudo reprogramar en este momento.' : 'Unable to reschedule right now.', 'error');
     });
 }
