+++ js/ui.js
@@ -2,20 +2,19 @@ import { withDeployAssetVersion } from './utils.js';
 import { loadDeferredModule, createWarmupRunner, bindWarmupTarget, scheduleDeferredTask, runDeferredModule } from './loader.js';
 import { closePaymentModal } from './booking.js';
 
-const UI_EFFECTS_URL = withDeployAssetVersion('/ui-effects.js?v=figo-ui-20260220-sync2');
-const MODAL_UX_ENGINE_URL = withDeployAssetVersion('/modal-ux-engine.js?v=figo-modal-ux-20260220-phase2-cachefix1');
+const UI_BUNDLE_URL = withDeployAssetVersion('/js/engines/ui-bundle.js');
 
 // UI Effects
 export function loadUiEffects() {
     return loadDeferredModule({
-        cacheKey: 'ui-effects',
-        src: UI_EFFECTS_URL,
-        scriptDataAttribute: 'data-ui-effects',
+        cacheKey: 'ui-bundle',
+        src: UI_BUNDLE_URL,
+        scriptDataAttribute: 'data-ui-bundle',
         resolveModule: () => window.PielUiEffects,
         isModuleReady: (module) => !!(module && typeof module.init === 'function'),
         onModuleReady: (module) => module.init(),
         missingApiError: 'ui-effects loaded without API',
-        loadError: 'No se pudo cargar ui-effects.js',
+        loadError: 'No se pudo cargar ui-effects (ui-bundle)',
         logLabel: 'UI effects'
     });
 }
@@ -51,14 +50,14 @@ function getModalUxEngineDeps() {
 
 export function loadModalUxEngine() {
     return loadDeferredModule({
-        cacheKey: 'modal-ux-engine',
-        src: MODAL_UX_ENGINE_URL,
-        scriptDataAttribute: 'data-modal-ux-engine',
+        cacheKey: 'ui-bundle',
+        src: UI_BUNDLE_URL,
+        scriptDataAttribute: 'data-ui-bundle',
         resolveModule: () => window.PielModalUxEngine,
         isModuleReady: (module) => !!(module && typeof module.init === 'function'),
         onModuleReady: (module) => module.init(getModalUxEngineDeps()),
         missingApiError: 'modal-ux-engine loaded without API',
-        loadError: 'No se pudo cargar modal-ux-engine.js',
+        loadError: 'No se pudo cargar modal-ux-engine (ui-bundle)',
         logLabel: 'Modal UX engine'
     });
 }
